[FE] refactor: slide 함수 로직 분리
함수 로직 긴 것들은
func 라는 디렉토리 밑에 분리해뒀다.
