+++ frontend/components/Slidebar/func/calculatePixels.ts
