+++ frontend/components/Slidebar/func/onNextClicked.ts
